registrar: account for ens fallback mechanism
The new ens registry has a fallback mechanism where the old registry will be
queried if an owner for a record cannot be found in the new registry.

The logic for this mechanism makes use to two sentinel values for the
`owner` field in a record:

- `address(0)`: no owner found in the new registry, try the old one
- `address(this)`: record is unowned, do not fallback to the old registrar

Calling `setOwner(record, address(0))` on the new registrar will in fact
set a value of `address(this)` for the owner field on that record.

the `recordExists` method only checks if the owner of a domain is
`address(0)` and returns `true` otherwise.

The combination of the above two points meant that revoked names (where
the owner had been set to `address(0)`) could not be reregistered.

This commit replaces the call to `recordExists` with a check based on
`ens.owner(node)` (which returns `address(0)` when the owner of a record
is `address(this)`), thus allowing reregistration of intentionally
revoked names.