registrar: remove redundant require
This check ensures that the owner of the "eth" tld in the ens registry
has been set. However the call to `setDomainOwner` will always fail if
this is not the case (due to the call to `transferFrom` later in the
method).

Removing as this is not a user facing method, and clarity of
implementation trumps clarity of error messages in this case.