registrar: setDomainOwner does not update resolver and ttl
The `setDomainOwner` method should not set the resolver and ttl values.
In particular the previous behaviour (setting the resolver to the
address of the new domain owner) seems almost certain to be wrong (and
cause disruption for downstream clients who may rely on the resolver for
the `radicle.eth` domain).