Registrar: unify commitment interfaces
This change introduces a more unified interface for the various commitment
methods, allowing us to drop two parameters from `commitBySigWithPermit`.

This does mean that `commitBySig` now reveals part of the preimage of
the commitment (`owner`), but this was already the case with the various
`permit` variants, and the `salt` means that getting the full preimage
remains computationally infeasible.